Replace deprecated std::auto_ptr<> with std::unique_ptr<>.
I also replaced a few bare pointers with std::unique_ptr<> for safety.

Note that Puzzle is copyable (and movable?), so in most places where a
std::unique_ptr<Puzzle> is returned or accepted, a plain Puzzle value
would also work, but I don't care enough about it to refactor it now.